Did AJAX and jQuery
App pretty much works now. Each new search replaces the html of the old
search and at the bottom there is a recent searches list. Next step is
to make recent searches clicklable but I think I will need to change
the function to take the artist name as an input.