Added bootstrap js for modal, added modal and did a ton of repositioning.
I read online that using absolute positioning is bad practice, so I
reworked a bunch of the css to fix that. It's looking good, but it had
a bit of white space so I threw in the modal (still not working) but I
included a JSFiddle with a working example of a modal in the comments
on the HTML doc for your use.